pyproject.toml
[project]
name = "lang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.3.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.14",
    "langchain-openai>=0.3.19",
    "langchain-tavily>=0.2.0",
    "langgraph>=0.4.7",
    "python-dotenv>=1.1.0",
]
